requirements.txt
tk
beautifulsoup4
lxml
reportlab
pillow
